refactor(youtube/open-links-externally): move to better package
