+++ src/main/kotlin/app/revanced/patches/youtube/misc/links/open/fingerprints/BindSessionServiceFingerprint.kt
@@ -1,17 +1,16 @@
-package app.revanced.patches.youtube.misc.openlinksexternally.fingerprints
+package app.revanced.patches.youtube.misc.links.open.fingerprints
 
 import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 import org.jf.dexlib2.AccessFlags
 import org.jf.dexlib2.Opcode
 
-object GetCustomTabPackageNameFingerprint : MethodFingerprint(
+object BindSessionServiceFingerprint : MethodFingerprint(
     returnType = "L",
-    access = AccessFlags.PUBLIC or AccessFlags.STATIC,
+    access = AccessFlags.PUBLIC or AccessFlags.FINAL,
     opcodes = listOf(
-        Opcode.CHECK_CAST,
+        Opcode.IPUT_OBJECT,
         Opcode.NEW_INSTANCE,
-        Opcode.INVOKE_DIRECT,
         Opcode.CONST_STRING
     ),
     strings = listOf("android.support.customtabs.action.CustomTabsService")
