+++ src/main/kotlin/app/revanced/patches/youtube/misc/links/open/fingerprints/GetCustomTabPackageNameFingerprint.kt
@@ -1,13 +1,13 @@
-package app.revanced.patches.youtube.misc.openlinksexternally.fingerprints
+package app.revanced.patches.youtube.misc.links.open.fingerprints
 
 import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 import org.jf.dexlib2.AccessFlags
 import org.jf.dexlib2.Opcode
 
-object InitializeCustomTabSupportFingerprint : MethodFingerprint(
-    returnType = "V",
-    access = AccessFlags.PUBLIC or AccessFlags.CONSTRUCTOR,
+object GetCustomTabPackageNameFingerprint : MethodFingerprint(
+    returnType = "L",
+    access = AccessFlags.PUBLIC or AccessFlags.STATIC,
     opcodes = listOf(
         Opcode.CHECK_CAST,
         Opcode.NEW_INSTANCE,
