+++ src/main/kotlin/app/revanced/patches/youtube/misc/links/open/fingerprints/InitializeCustomTabSupportFingerprint.kt
@@ -1,16 +1,17 @@
-package app.revanced.patches.youtube.misc.openlinksexternally.fingerprints
+package app.revanced.patches.youtube.misc.links.open.fingerprints
 
 import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 import org.jf.dexlib2.AccessFlags
 import org.jf.dexlib2.Opcode
 
-object BindSessionServiceFingerprint : MethodFingerprint(
-    returnType = "L",
-    access = AccessFlags.PUBLIC or AccessFlags.FINAL,
+object InitializeCustomTabSupportFingerprint : MethodFingerprint(
+    returnType = "V",
+    access = AccessFlags.PUBLIC or AccessFlags.CONSTRUCTOR,
     opcodes = listOf(
-        Opcode.IPUT_OBJECT,
+        Opcode.CHECK_CAST,
         Opcode.NEW_INSTANCE,
+        Opcode.INVOKE_DIRECT,
         Opcode.CONST_STRING
     ),
     strings = listOf("android.support.customtabs.action.CustomTabsService")
